#178: Customize GlobalLayout for better header color scheme control
